移植 getPastYears
